% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_summary.R
\name{glm_summary}
\alias{glm_summary}
\title{Summary statistics for Linear Models}
\usage{
glm_summary(object, ..., print = TRUE)
}
\arguments{
\item{object}{svyglm result object.}

\item{...}{Any further parameters.}

\item{print}{Logical, if summary table should be printed.}
}
\value{
Dataframe with summary staistics
}
\description{
Summary statistics for Linear Models
}
\examples{
glm_summary(mitools::MIcombine(results$deu))   # Results for German dataset
glm_summary(mitools::MIcombine(results[[2]]))  # Results for second dataset
}
