% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piaac_glm.R
\name{piaac_glm}
\alias{piaac_glm}
\title{Linear Regression with PIAAC data}
\usage{
piaac_glm(datapath, countries, variables, formula, FUN = piaac_tidy, ...)
}
\arguments{
\item{datapath}{Path to the PIAAC data. A string.}

\item{countries}{Vector with three-letter country codes (ISO3166-1 alpha-3).}

\item{variables}{Vector with variable names.}

\item{formula}{Formula for the linear model.}

\item{FUN}{Function name for data tidying.}

\item{...}{Any parameters forwarded to FUN.}
}
\value{
List of regression results. Individual list elements are to be handled with mitools.
}
\description{
Linear Regression with PIAAC data
}
\examples{
results <- piaac_glm("~/data/", c("deu", "usa"), c("gender"), pvlit ~ gender)
results <- piaac_glm("~/data/", c("deu", "usa"), c("gender"), pvlit ~ gender, FUN = tidy_data)
}
