% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piaac_read.R
\name{piaac_read}
\alias{piaac_read}
\title{Reads PIAAC data}
\usage{
piaac_read(datapath, cntry)
}
\arguments{
\item{datapath}{Path to the PIAAC data. A string.}

\item{cntry}{Three-letter country code (ISO3166-1 alpha-3). A string.}
}
\value{
A tibble, imported by the haven package.
}
\description{
Reads PIAAC data
}
\examples{
d <- piaac_read("~/data/", "usa")
}
