% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piaac_tidy.R
\name{piaac_tidy}
\alias{piaac_tidy}
\title{Placeholder function for data tidying}
\usage{
piaac_tidy(d, variables, ...)
}
\arguments{
\item{d}{Dataframe or Tibble to tidy}

\item{variables}{Vector of variable names specified in model function call.}

\item{...}{Any parameters specified in model function call.}
}
\value{
Dataframe or Tibble
}
\description{
Placeholder function for data tidying
}
